Release/1.0.0

+++ src/megacmdshell/megacmdshellcommunicationsnamedpipes.cpp
@@ -42,17 +42,6 @@
 #define _O_U8TEXT 0x00040000
 #endif
 
-#if defined(_WIN32) && !defined(WINDOWS_PHONE)
-#include "mega/thread/win32thread.h"
-class MegaThread : public mega::Win32Thread {};
-#elif defined(USE_CPPTHREAD)
-#include "mega/thread/cppthread.h"
-class MegaThread : public mega::CppThread {};
-#else
-#include "mega/thread/posixthread.h"
-class MegaThread : public mega::PosixThread {};
-#endif
-
 bool MegaCmdShellCommunicationsNamedPipes::confirmResponse; //TODO: do all this only in parent class
 bool MegaCmdShellCommunicationsNamedPipes::stopListener;
 mega::Thread *MegaCmdShellCommunicationsNamedPipes::listenerThread;
@@ -261,11 +250,11 @@ bool MegaCmdShellCommunicationsNamedPipes::isFileOwnerCurrentUser(HANDLE hFile)
         }
         else
         {
-            wcerr << L"Unmatched owner - current user -> " << AcctName << L" - " << username;
+            std::wcerr << L"Unmatched owner - current user -> " << AcctName << L" - " << username;
 #ifndef __MINGW32__
-            wcerr << L" IsUserAdmin=" << IsUserAnAdmin();
+            std::wcerr << L" IsUserAdmin=" << IsUserAnAdmin();
 #endif
-            wcerr << L" SIDOwner=" << stringSIDOwner << endl;
+            std::wcerr << L" SIDOwner=" << stringSIDOwner << endl;
             return false;
         }
 
@@ -472,9 +461,9 @@ MegaCmdShellCommunicationsNamedPipes::MegaCmdShellCommunicationsNamedPipes()
     listenerThread = NULL;
 }
 
-int MegaCmdShellCommunicationsNamedPipes::executeCommandW(wstring wcommand, int (*readconfirmationloop)(const char *), OUTSTREAMTYPE &output, bool interactiveshell)
+int MegaCmdShellCommunicationsNamedPipes::executeCommandW(wstring wcommand, string (*readresponse)(const char *), OUTSTREAMTYPE &output, bool interactiveshell)
 {
-    return executeCommand("", readconfirmationloop, output, interactiveshell, wcommand);
+    return executeCommand("", readresponse, output, interactiveshell, wcommand);
 }
 
 /**
@@ -512,7 +501,7 @@ bool outputtobinaryorconsole(void)
     return false;
 }
 
-int MegaCmdShellCommunicationsNamedPipes::executeCommand(string command, int (*readconfirmationloop)(const char *), OUTSTREAMTYPE &output, bool interactiveshell, wstring wcommand)
+int MegaCmdShellCommunicationsNamedPipes::executeCommand(string command, std::string (*readresponse)(const char *), OUTSTREAMTYPE &output, bool interactiveshell, wstring wcommand)
 {
     HANDLE theNamedPipe = createNamedPipe(0,command.compare(0,4,"exit")
                                           && command.compare(0,4,"quit")
@@ -568,7 +557,7 @@ int MegaCmdShellCommunicationsNamedPipes::executeCommand(string command, int (*r
         return -1;
     }
 
-    while (outcode == MCMD_REQCONFIRM)
+    while (outcode == MCMD_REQCONFIRM || outcode == MCMD_REQSTRING )
     {
         int BUFFERSIZE = 1024;
         string confirmQuestion;
@@ -585,18 +574,39 @@ int MegaCmdShellCommunicationsNamedPipes::executeCommand(string command, int (*r
             cerr << "ERROR reading confirm question: " << ERRNO << endl;
         }
 
-        int response = MCMDCONFIRM_NO;
-
-        if (readconfirmationloop != NULL)
+        if (outcode == MCMD_REQCONFIRM)
         {
-            response = readconfirmationloop(confirmQuestion.c_str());
+            int response = MCMDCONFIRM_NO;
+
+            if (readresponse != NULL)
+            {
+                response = readconfirmationloop(confirmQuestion.c_str(), readresponse);
+            }
+
+            if (!WriteFile(newNamedPipe, (const char *) &response, sizeof(response), &n, NULL))
+            {
+                cerr << "ERROR writing confirm response to namedPipe: " << ERRNO << endl;
+                return -1;
+            }
+        }
+        else // MCMD_REQSTRING
+        {
+            string response = "FAILED";
+
+            if (readresponse != NULL)
+            {
+                response = readresponse(confirmQuestion.c_str());
+            }
+
+            wstring wresponse;
+            stringtolocalw(response.c_str(),&wresponse);
+            if (!WriteFile(newNamedPipe, (char *) wresponse.data(), DWORD(wcslen(wresponse.c_str())*sizeof(wchar_t)), &n, NULL))
+            {
+                cerr << "ERROR writing confirm response to namedPipe: " << ERRNO << endl;
+                return -1;
+            }
         }
 
-        if (!WriteFile(newNamedPipe, (const char *) &response, sizeof(response), &n, NULL))
-        {
-            cerr << "ERROR writing confirm response to namedPipe: " << ERRNO << endl;
-            return -1;
-        }
 
         if (!ReadFile(newNamedPipe, (char *)&outcode, sizeof(outcode),&n, NULL))
         {
